backend/flask_app/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==38.0.3  # Replace with a stable version
decorator==5.1.1
dnspython==2.2.1  # Replace with a stable version
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Session==0.8.0
frozenlist==1.5.0
genai==2.1.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
msgspec==0.18.6
multidict==6.1.0
openai==0.27.10
packaging==24.2
parso==0.8.4
pdf2image==1.17.0
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pymongo==4.1.1  # Replace with a stable version
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
regex==2024.11.6
requests==2.32.3
rsa==4.9
six==1.17.0
stack-data==0.6.3
tabulate==0.9.0
tiktoken==0.3.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
